% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_instance.R
\name{calc_instance}
\alias{calc_instance}
\title{Calcultes the number of instances}
\usage{
calc_instance(ninstances = NULL, cpower = NULL, d = NULL, algorithms,
  alpha = 0.05, alpha.correction = "holm", compare.type = "one",
  direction = 2, nmax = Inf)
}
\arguments{
\item{ninstances}{the number of instances to be used in the experiment.
See \code{Number of Instances} for details.}

\item{cpower}{power deisred for the comparision
See \code{Number of Instances} for details.}

\item{d}{desired effect size normalized by the standart deviation
See \code{Number of Instances} for details.}

\item{algorithms}{the number of algorithms to be comparred or a list object
containing lists defining all algorithms to
be used in the experiment}

\item{alpha}{significance level for the confidence intervals on the means of
each algo-problem pair.}

\item{alpha.correction}{is the type of alpha correction used - default is
\code{holm}}

\item{compare.type}{defines if the comparition is 'all vc all' or 'one vs all'-
default is \code{one}}

\item{direction}{is the direction of the distribution 1 for 1-sided and 2 for
2-sided - default is \code{2}}

\item{nmax}{maximum allowed number of instances}
}
\value{
a list object containing the following items:
\itemize{
   \item \code{ninstances} - number of instances
   \item \code{cpower} - the power of the comparision
   \item \code{d} - the effect size
   \item \code{corrected.aplha} - the corrected alpha
   and some input parameters: algorithms, alpha and nmax as given
}
}
\description{
Return either the number of isntances, the effect size or the power of the
comparision considering the other two
}
\details{
This routine uses the closed formula of the t-test to calculates the number
of instances necessary for a comparition of k algorithms, considering given
power and  effect size.
In case the number of instaces is already known, it can return the effect
size or power, given the other
}
\section{Number of Instances}{

}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}), Fernanda Takahashi (\email{fernandact@ufmg.br})
}

