% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_n.R
\name{calc_n}
\alias{calc_n}
\title{Sample size for fixed-width confidence interval for the mean}
\usage{
calc_n(delta, sd, alpha = 0.05)
}
\arguments{
\item{delta}{Desired halfwidth of the confidence interval}

\item{sd}{Estimated standard deviation}

\item{alpha}{Significance level of the CI.}
}
\value{
The required sample size for obtaining a (1 - \code{alpha})-CI with
   halfwidth \code{delta}
}
\description{
Number of observations for obtaining a CI for the mean with predefined
   halfwidth
}
\details{
\code{calc_n} calculates the required number of observations needed to
   generate a fixed-width (1 - \code{alpha})-CI for the mean based on the t
   distribution.
}
\examples{
calc_n(delta = 3, sd = 8)
calc_n(delta = .25, sd = 1, alpha = 0.01)
}
\author{
Felipe Campelo
}

