% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/envelope.R
\name{envelope}
\alias{envelope}
\title{Envelope Routine}
\usage{
envelope(Alg, Prob, e, n0, a)
}
\arguments{
\item{Alg}{list of algorithms - list()}

\item{Prob}{list of problems, as lists of parameters of Alg - list(list())}

\item{e}{desired standard error - positive numeric}

\item{n0}{initial sample size - positive integer}

\item{a}{desired significance level (alpha) - positive numeric}
}
\value{
list with \code{list(db.full, db.short)} containig a database with all the generated data and a collapsed database with the sample size, mean and standart deviation - list of databases
}
\description{
Call the sampling routine for a given set o algorithms and problems
}

