% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval_power.R
\name{eval_power}
\alias{eval_power}
\title{Post-experimental power analysis}
\usage{
eval_power(data, alpha = 0.05, delta = NA, power = NA,
  alternative = "two.sided")
}
\arguments{
\item{x}{}
}
\value{
y
}
\description{
Power analysis of the comparison of mean performance of two algorithms on N
problem instances.
}
\details{
The details of this routine come here...
}
\author{
Felipe Campelo
}

