% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_nreps.R
\name{run_nreps}
\alias{run_nreps}
\title{Automatically determine sample size for an algorithm on a problem instance}
\usage{
run_nreps(problem, algo, dmax, method = c("parametric", "bootstrap"),
  alpha = 0.05, nstart = 25, nmax = Inf, seed = NULL, nboot = 1000,
  ncpus = 1)
}
\arguments{
\item{problem}{a list object containing the definitions of the problem
instance. See \code{\link{run_chase}} for details.}

\item{algo}{a list object containing the definitions of the algorithm See
\code{\link{run_chase}} for details.}

\item{dmax}{desired confidence interval halfwidth for the estimated mean
performance of algorithm \code{algo} on instance \code{problem}.}

\item{method}{method used to calculate the confidence interval. Currently
supported methods are "parametric" and "bootstrap"}

\item{alpha}{significance level for the confidence intervals.}

\item{nstart}{initial number of algorithm runs.
See \code{\link{run_chase}} for details.}

\item{nmax}{maximum allowed sample size.
Defaults to \code{nmax = Inf}.}

\item{seed}{seed for the random number generator.
Defaults to NULL. See \code{\link{run_chase}} for details.}

\item{nboot}{number of bootstrap samples.}

\item{ncpus}{number of cores to use for bootstrap.}
}
\value{
a list object containing the following items:
\itemize{
   \item \code{observations} - vector of observed performance values
   \item \code{xbar} - sample mean of the performance of \code{algo} on \code{problem}
   \item \code{se} - standard error of the mean
   \item \code{n} - number of observations
   \item \code{seed} - the seed used for the PRNG
}
}
\description{
Iteratively calculates the required sample size for an algorithm on a
   problem, so that the final number of replicates yields an estimation of
   mean performance with a fixed uncertainty.
}
\details{
The details of this routine come here...
}
\author{
Felipe Campelo
}
\seealso{
run_chase
}

