% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampling}
\alias{sampling}
\title{Sampling Routine}
\usage{
sampling(A, P, e = 0.05, n0 = 10, a = 0.1)
}
\arguments{
\item{A}{name of the function (algorithm) to be evaluated}

\item{P}{parameters  of A (problem to be evaluated)}

\item{e}{desired standard error}

\item{n0}{initial sample size}

\item{a}{desired significance level (alpha)}
}
\value{
vector \code{values} containing the sampled
}
\description{
Samples the number of executions needed to obtain the desired error
}
\section{References}{

Botella, J., Ximenez, C., Revuelta, J. and Suero, M.,
"Optimization of sample size in controlled experiments: the CLAST rule.",
Behavior Research Methods, 38(1) 65 - 76, 2006
}

