% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ci.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculates confidence interval for a specified statistic}
\usage{
calc_ci(x, FUN = "mean", alpha = 0.05, method = "parametric",
  nboot = 1000, ncpus = 4)
}
\arguments{
\item{x}{vector of observations}

\item{FUN}{name of statistic function
Defaults to \code{FUN = "mean"}.}

\item{alpha}{desired significance level for the interval.
Defaults to \code{alpha = 0.05}.}

\item{method}{method used to calculate the interval.
Defaults to \code{method = "parametric"}.}

\item{nboot}{number of bootstrap resamples (if \code{method} = "bootstrap").
Defaults to \code{nboot = 1000}.}

\item{ncpus}{number of cores to use for bootstrap
(if \code{method} = "bootstrap"). Defaults to \code{ncpus = 4}.}
}
\value{
a list object containing the following items:
\itemize{
   \item \code{ci} - numeric vector of length 2 containing the confidence
     interval
   \item \code{se} - standard error of the base statistic
}
}
\description{
Calculates the confidence interval for a statistic, based on a given sample.
}
\details{
The details of this routine come here...
(s.e. for median is based on asymptotic formula for normal
distribution/large sample size, see
R.R. Sokal, F.J. Rohlf, "Biometry", W.H. Freeman, 4th ed., 2011, p. 139)
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}), Fernanda Takahashi (\email{fernandact@ufmg.br})
}

