% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_instance.R
\name{calc_instance}
\alias{calc_instance}
\title{Power calculations for the CHASE procedure}
\usage{
calc_instance(N = NULL, cpower = NULL, d = NULL, algorithms,
  alpha = 0.05, alpha.correction = "holm", comparetype = "one",
  direction = 2, nmax = Inf)
}
\arguments{
\item{N}{the number of instances to be used in the experiment.
See \code{Number of Instances} for details.}

\item{cpower}{power deisred for the comparision
See \code{Number of Instances} for details.}

\item{d}{desired effect size normalized by the standart deviation
See \code{Number of Instances} for details.}

\item{algorithms}{the number of algorithms to be comparred or a list object
containing lists defining all algorithms to
be used in the experiment}

\item{alpha}{significance level for the confidence intervals on the means of
each algo-problem pair.}

\item{alpha.correction}{is the type of alpha correction used - default is
\code{holm}}

\item{comparetype}{defines if the comparition is 'all vc all' or 'one vs all'-
default is \code{one}}

\item{direction}{is the direction of the distribution 1 for 1-sided and 2 for
2-sided - default is \code{2}}

\item{nmax}{maximum allowed number of instances}
}
\value{
a list object containing the following items:
\itemize{
   \item \code{N} - number of instances
   \item \code{cpower} - the power of the comparision
   \item \code{d} - the effect size
   \item \code{corrected.alpha} - the corrected alpha
}
The output list also includes some of the input parameters, for convenience:
algorithms, alpha and nmax as given
}
\description{
Compute the number of instances required for a given comparison of algorithms,
or determine the power curve of the comparision.
}
\details{
This routine calculates the number of instances necessary for a comparison
of \code{K} algorithms, considering a desired power level and a given value
for the minimally interesting effect size (either standardized in terms of
Cohen's d coefficient or given as an absolute value, in which case an
estimate for the residual standard deviation must be provided).

If the number of instaces is predefined (e.g., when using standard benchmark
sets), this routine can return the power for a given effect size, or the
effect size that can be detected with a certain power.
}
\section{Usage}{


This routine can be used in several ways. Five parameters are involved in
power calculations for statistical comparisons of algorithms:

\itemize{
 \item the significance level (\code{alpha})
 \item the number of instances (\code{N})
 \item the power of the comparison (\code{cpower})
 \item the magnitude of the actual difference (\code{delta})
 \item the residual standard deviation (\code{sigma})
}

The two last items in this list can often be combined into a standardized
effect size \code{d = delta/sigma}, that is, in a non-dimensional effect size
indicator that is given as the number of standard deviations.

Depending on which inputs are provided to the routine, it performs different
calculations. The most common usages are listed below:

\strong{Case 1:} \code{N = NULL}, with given \code{d}: calculates the
required number of instances, \code{N}

\strong{Case 2:} \code{N = NULL}, with given \code{delta} and \code{sigma}:
since \code{d = delta / sigma}, this is equivalent to \strong{Case 1})

\strong{Case 3:} \code{d = delta = sigma = NULL}: calculates the smallest
standardized effect size \code{d} that can be detected with power
\code{cpower}.

\strong{Case 4:} \code{cpower = NULL}: calculates the power of the comparison
to detect a standardized effect size of \code{d} or greater.

\strong{Case 5:} \code{d = cpower = NULL}: calculates the power curves
\code{d X cpower} of the comparison.

\strong{Case 6:} \code{N = cpower = NULL}: calculates the power curves
\code{N X cpower} of the comparison.

\strong{Case 7:} \code{N = d = NULL}: calculates the power curves
\code{N X d} of the comparison.\eqn{a \neq b}
}

\section{MHT correction}{

}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}),
         Fernanda Takahashi (\email{fernandact@ufmg.br})
}

