% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_instance.R
\name{calc_instance}
\alias{calc_instance}
\title{Calcultes the number os instances}
\usage{
calc_instance(ninstances = NULL, cpower = NULL, d = NULL,
  comparetype = "one", direction = 2, algorithms, alpha = 0.05,
  alpha.correction = "holm", nstart = 10, nmax = Inf, seed = NULL)
}
\description{
param
}

